- Move RetrofitClassAware... builder method to the RetrofitMeterIdPrefixFunction
